Added Column.SetSortType
When the column is type of decimal, date, currency etc, it was not
sorting correctly because the default sort type is text. Added a method
to set the sort type on column.

Here are the possible sorttype from jqgrid
(http://www.trirand.net/documentation/php/_2v70w5p2k.htm)

Used when datatype is local. Defines the type of the column for
appropriate sorting.Possible values:
int/integer - for sorting integer
float/number/currency - for sorting decimal numbers
date - for sorting date
text - for text sorting. The script automaticall fill this property.